extract generation functions
